Fix baremetal serial console autoenable
If ironic_serial_console_autoenable is set to true, ironic compute node
serial consoles are automatically enabled during kayobe overcloud post
configure. However, the baremetal-compute-serial-console.yml playbook
expects a variable called cmd to be defined, and set to 'enable' in
order to enable serial consoles. This patch fixes the issue.

We also add a conditional to only import the playbook when ironic is
enabled in the overcloud.

Change-Id: Ice26a113f9023b1a25a4af0552d07c5e95e64556
Story: 2006662
Task: 36924