Fix validation of locally built IPA when ironic is disabled
In an environment where ironic is disabled, but IPA images are built
locally for the seed, the following task will fail in a number of
commands, e.g. kayobe overcloud service deploy:

Validate the presence of locally built Ironic Python Agent (IPA) images

We need to only execute this task when ironic is enabled.

Change-Id: I61067164a003bcdf6079cdaca36f0e1755186f60
Story: 2006669
Task: 36935