Fix kolla ansible virtualenv creation on the seed
Backport: Stein, Rocky, Queens

In https://review.opendev.org/#/c/671502/ kolla ansible virtualenv
creation was modified to only happen on hosts in the kolla-ansible
inventory. However, that change omitted the seed. It passed CI because
the seed and overcloud are both in the inventory.

This change adds the seed group to the inventory_hostname lookup.

Change-Id: Ib5fb726600debee5df1f7085ea37b583bfccefcf
Story: 2006247
Task: 37368