Fix kayobe overcloud provision without seed in limit
If you run 'kayobe overcloud provision' for the first time on a system,
and do not include the seed in the limit (e.g. --limit overcloud), then
the following error is seen:

    Destination directory /etc/kolla/bifrost/inventory/host_vars does
    not exist

This is because the first play in ansible/kolla-bifrost-hostvars.yml
executes against the seed. This is not ideal, as the --limit argment is
useful with this command, as a way to provision specific nodes.

This change fixes the issue by moving the initialisation tasks to the
second play which is targeted at the overcloud group. The run_once
parameter ensures it happens once.

Change-Id: I7dff72a50f0592a3564c9d84d684db562c30626a
Story: 2006757
Task: 37247
(cherry picked from commit e0ff12eaef5f8eb00c4f044196bdc5650110dd07)