Fix IPA builds without packages specified
Since story 2006855, it is possible to specify a list of packages when
building IPA images. However, this introduced an issue in the case where
no packages are specified. See
https://github.com/stackhpc/ansible-role-os-images/issues/26.

This change fixes the issue by updating stackhpc.os-images role to
v1.7.0, which has a fix for this issue.

Change-Id: I70d20381bec3f5b92bb467e6c506e32bb13548c2
Story: 2007069
Task: 37952