Allow OVS bridges to connect directly to interface
Currently we require a Linux bridge to exist between OVS and the
physical interface. This is necessary if you want to set an IP on the
native VLAN of that interface, but that is not always the case.

This change allows the physical interface (or any non-bridge interface)
to be plugged into OVS.

Change-Id: I2172a74f4719605f6ec81fadec46ce49f8310a96
Story: 2007364
Task: 38920