CI: fix kayobe-tox-molecule job
The kayobe-tox-molecule job is currently failing on stable/train and
earlier branches. This was caused by the 13.0 release of the 'sh'
package [1]. A 13.1 release fixed this for Python 3, but there is no fix
available for Python 2 currently.

This change works around the issue by pinning 'sh' to <13.0 in
test-requirements.txt.

[1] https://github.com/ansible-community/molecule/issues/2676

Change-Id: I0df1dadb3b48736cdbd25112517e524533550210