Remove newline from default public_net_name
The public network name is set via public_net_name, and its default
value is the first item in the external_net_names list. The variable
uses the YAML folded block scalar notation, but does not use the
chomping indictator, which means the string is assigned a trailing
newline.

This change fixes the issue by adding the chomping indicator.

Change-Id: I68fb65cd3bc7c4390813747f9ac4ecebc3b4145f
Story: 2007654
Task: 39731