Fix ironic inspector store endpoint configuration
When performing inspection on baremetal compute nodes, inspector
currently has the following failure when storing inspection data to the
introspection store:

    No 'keystoneauth1.plugin' driver found, looking for 'token_endpoint'

This happens because the token_endpoint auth type is no longer
available. The new way to do this is to set auth_type = none and provide
an endpoint override.

Story: 2007795
Task: 40049

Change-Id: Ideada0f299f34d16d1286633b6e21404b736918b