CI: don't run virtualenv on an existing virtualenv
The seed upgrade job is currently failing on CentOS 7:

Traceback (most recent call last):
  File "/usr/lib/python2.7/site-packages/virtualenv.py", line 2327, in <module>
    main()
  File "/usr/lib/python2.7/site-packages/virtualenv.py", line 712, in main
    symlink=options.symlink)
  File "/usr/lib/python2.7/site-packages/virtualenv.py", line 924, in create_environment
    site_packages=site_packages, clear=clear, symlink=symlink))
  File "/usr/lib/python2.7/site-packages/virtualenv.py", line 1369, in install_python
    os.symlink(py_executable_base, full_pth)
OSError: [Errno 17] File exists

This is likely due to changes in the nodepool infra image which removed
the default installation of virtualenv.

This change avoids running virtualenv on an existing virtualenv.

It also installs the python six package in CI, which is required by
Tenks.

Depends-On: https://review.opendev.org/736280

Change-Id: Id482ae69fc84b106d9d11a9a1693d94ce5f34ba5