Pin ansible-lint to 4.2.0
In stable branches we don't want to follow the latest rules and have
constantly broken CI but rather ensure we control quality checks and act
as needed.

Most other test requirements are actually pinned via upper-constraints,
but ansible-lint is not in global-requirements so it is entirely up to
us to handle pinning.

Change-Id: I22057112ff06cb1aec081dfd05b984f048ce0594
Co-Authored-By: Radosław Piliszek <radoslaw.piliszek@gmail.com>