Fix overcloud provisioning
Affects master & Victoria only.

Since Victoria, Bifrost added a play to gather facts for localhost at
the beginning of deploy-dynamic.yaml. Kayobe executes the playbook with
a limit, which does not include localhost. This causes fact gathering to
be skipped for localhost:

    PLAY [Gather facts on localhost] ***********************************
      skipping: no hosts matched

This change fixes the issue by adding localhost to the limit.

Change-Id: I1264832398744c48a6f706efe91f084890a37772
Story: 2008388
Task: 41311