Fix overcloud host image resolv workaround on CentOS 8.3
If a seed's bifrost_deploy container is deployed using a CentOS 8.2 base
image, prior to the CentOS 8.3 release, then a user runs 'kayobe seed
service deploy', the following task may fail:

    TASK [Ensure the overcloud host image has bogus name server entries removed]

This seems to be an issue with libgcrypt not being upgraded with
qemu-kvm, and becoming incompatible. The workaround is to ensure
libgcrypt is upgraded also.

Story: 2008430
Task: 41379

Change-Id: I82313fcd258e58969e0b68ae91a78943df537a7f