Ubuntu: fix pip configuration for kolla user
On Ubuntu, Kayobe can fail creating pip configuration for the kolla
user, with the following error:

    Failed to set permissions on the temporary files Ansible needs to
    create when becoming an unprivileged user

An earlier error seen with Ansible verbose mode shows that the setfacl
command is missing. It can be installed using the acl package, however
this is not always present in Ubuntu 20.04.

This change installs the acl package in the pip role.

Story: 2004960
Task: 41492

Change-Id: Ib5462684fa5fd307068c67f9a13f73b80967beb4