Fix deployment image build in Victoria
In the Victoria release, deployment images fail to build with the
following error:

/opt/kayobe/venvs/ipa-build-dib/lib/python3.6/site-packages/diskimage_builder/lib/common-functions: line 343: element-info: command not found

This is caused by the virtualenv bin directory not being included in
PATH while executing code from common-functions, which results in the
element-info script not being found.

This doesn't happen in Ussuri because older versions of DIB were
activating the virtualenv automatically [1].

This doesn't happen in master because its upper constraints allow a
newer DIB which doesn't use element-info anymore.

The fix is to activate the DIB virtualenv, which is done since
stackhpc.os-images v1.10.1 [2]. Bump in master first then backport to
stable/victoria.

[1] https://review.opendev.org/c/openstack/diskimage-builder/+/734414
[2] https://github.com/stackhpc/ansible-role-os-images/releases/tag/v1.10.1

Change-Id: I5efdbd52556721914fe69d7c6ba454b2c721b643
Story: 2008550
Task: 41673