CI: retry overcloud jobs that land on vexxhost clouds
Jobs that test baremetal deployment are currently failing when run on
Vexxhost infrastructure. The seed VM job also has trouble on Vexxhost,
although does sometimes pass. It's not clear why yet, and we are still
investigating. Until we have a solution, fail in the pre phase when we
land on vexxhost clouds, in order to trigger the zuul retry mechanism.

Also use the normal number (5) of attempts for upgrade jobs, since these
no longer deploy the previous release in the pre phase.

Finally, incorporate a fix from
https://review.opendev.org/c/openstack/kayobe-config-dev/+/777076 to
allow this to merge. This part can be reverted once the
kayobe-config-dev patch merges.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2021-February/020685.html

Change-Id: Ie5ca04428855cbaf505adc57de76ea0b2d404080