Ubuntu: refactor network.yml playbook
The network.yml playbook is reduced in complexity, now performing
validation before including a role specific to the OS family of the host
(network-redhat or network-debian). This will allow the network
configuration to diverge between OS families without adding excessive
conditionals.  Currently the content of the roles is identical.

Story: 2004960

Change-Id: Iaa69e5303e554e6b30a23a3e84e36d09972888c9