Ubuntu: define implied VLAN parent interfaces in networkd
On CentOS, when using the MichaelRigart.interfaces role, defining a VLAN
interface implies the existence of its parent interface. This is handled
via the CentOS network interface scripts. With systemd-networkd, the
parent interface is not implicit, and must be configured. The parent
interface network configuration file needs to contain this:

    [Network]
    VLAN = eth0.42

This change adds support for these implied parent interfaces. It also
removes the Broadcast = true option, which was added in an attempt to
fix an issue but is not supported in the [Network] section, and
generates a warning:

    Unknown key name 'Broadcast' in section 'Network', ignoring.

Story: 2004960
Task: 42368
Task: 42367

Change-Id: I5d105006fad6e7e80473b9d9fa693de644c35a6d