Ubuntu: avoid shadowing of networkd config files
Kayobe removes any systemd-networkd configuration generated by netplan,
since it can shadow our own configuration. This is performed via a
handler, but in practice it may run after the handler that restarts
systemd-networkd. This can leave the system using the netplan
configuration, which might not be correct. For example, if the interface
should be a bridge port.

This change fixes the issue by flushing handlers before importing the
systemd-networkd role.

Change-Id: I62ef7746aca4c6889526692cd0897e422937e70e
Story: 2004960
Task: 42415