Make docker registry network mode configurable
Adds a new flag, 'docker_registry_network_mode', which defaults to
'host'. This may be used to set the network mode of the Docker registry
container.

This is a follow up to I404dd52701426a10c2e92727bd52b7fd7112abf6, which
changed the network mode from the default of bridge to host. It allows
that change to be backported to stable branches, without modifying the
default value.

Change-Id: Ic8ec3bb98f8f016e1d089bf10bd0538264394241