Fixes for Ansible 2.10
os_image_facts renamed to os_image_info, and no longer sets a fact.
Instead we use the registered result. The same is true for
os_networks_facts and os_networks_info.

We did not catch the networks issue in CI because the code only runs
with a VLAN provisioning network, and we use a flat network.

Change-Id: I6a801e8bb5171bebd2b3f20d6140329d38ea36a5