Fix overcloud post configure Python interpreter issues
Some time ago in Kayobe we switched from using set_fact to override the
ansible_python_interpreter variable to using a task variable. However,
not all of the galaxy roles were updated to match. This can lead to an
issue where a fact set by the os_networks role overrides the task
variable set in Kayobe.

Note that we haven't seen this in CI since the code path requires the
use of a VLAN provisioning network, and CI uses a flat network.

Change-Id: I14289df5e248b5b47fb8e60c6eea1ac845f50fd3
Story: 2008284
Task: 41158