Remove default time sources from customisable configuration
It seems unlikely that anyone would want to extend the time servers
variable, so let's keep it simple and not add an additional variable
where one isn't necessary. This is also consistent with how we configure
DNS.

These variables haven't made it into an official release yet, so now
is the time to clean up.

Change-Id: I3c15c6976296abf9c78f1480c1779b854145c7ca