chrony: cleanup during overcloud host upgrade
The Kolla Ansible chrony container is disabled by default in the Wallaby
release. A new kolla-ansible chrony-cleanup command can be used to clean
up the container.

This change extends the 'kayobe overcloud host upgrade' command to cover
cleaning up the chrony container (if disabled) and deploying a host
chrony daemon.

Depends-On: https://review.opendev.org/c/openstack/kolla-ansible/+/792119

Change-Id: I275102ec6b5bab6982577b52fd29654c874446ce