ntp: check for ntp group in inventory
The Wallaby release adds support for configuring a chrony daemon on
hosts in the ntp group. The kolla chrony container is disabled by
default, meaning that users could be left without an NTP client if they
forget to add hosts to the ntp group.

This change adds a check for the existence of an ntp group in the
inventory. The group may be empty.

Change-Id: Ic0ff71bb11692eb6c5699a1673df2d16b8f894ec
Story: 2007872
Task: 42533
(cherry picked from commit 9d3045dbf0b6a45cd9c2dcb0a5fae8c8d81045f7)