Revert "CI: switch to quay.io for container images"
Kolla container images for stable/train are not being updated on
quay.io. We need a new bifrost-deploy image with the systemd package
from CentOS 8.4 in order to fix the following error when deploying on a
CentOS 8.4 host:

    Could not find the requested service mariadb: host

Switch back to images from Docker Hub, which are still updated weekly.
An image with CentOS 8.4 packages should be published on 2021-06-13.

This reverts commit b26f4c1cdf9a70db267b73535e06b882a07917d0.

Change-Id: Ib91fda6420a8255ba5d3cbb7de98613c492eb9b7