CI: Revert to Docker Hub images for upgrade jobs
Kolla container images for stable/train are not being updated on
quay.io. We need a new bifrost-deploy image with the systemd package
from CentOS 8.4 in order to fix the following error when deploying on a
CentOS 8.4 host:

    Could not find the requested service mariadb: host

Switch back to images from Docker Hub, which are still updated weekly,
for the upgrade jobs. An image with CentOS 8.4 packages should be
published on 2021-06-13.

Updating kayobe-overcloud-upgrade-base wasn't required to fix CI issues,
but it allows us to test upgrades from the latest train images.

Note that unlike other jobs, this makes the upgrade jobs use images
updated weekly for ussuri as well, which should not be an issue.

This partially reverts commit 0059a8cfa8e981f63c72724128d0d4749b4b1a57.

Change-Id: I233b61209433f89f4a46354d081a038b24cd2800